minor refactoring and fixes
